Avoid modifying the address factory
We'll be changing the frontend of the extension to
SolidusStarterFrontend in the following commits. In the resulting
update, specs for this extension and specs for the frontend will
be run together in a single dummy app. Based on my tests, it seems
the factory modifications are causing the coupon code spec from the
starter frontend spec suite to break.

Instead of modifying the address factory, we're taking the route of
creating a subclass of the factory that is namespaced to
solidus_braintree. This makes it safer to merge the extension's specs
with the starter frontend specs.

Done:

* Created solidus_braintree_address factory.
* Created with_first_and_last_name trait.
* Created with_fixed_zipcode trait.
* Used with_fixed_zipcode trait in OrderWalkthrough class.